+++ types/lib/b2c/client.d.ts
@@ -1,5 +1,3 @@
-/// <reference types="node" />
-import * as http from "http";
 import { CryptoWallets } from "./crypto_wallets";
 import { MagicLinks } from "./magic_links";
 import { OAuth } from "./oauth";
@@ -9,14 +7,8 @@ import { Sessions } from "./sessions";
 import { TOTPs } from "./totps";
 import { Users } from "./users";
 import { WebAuthn } from "./webauthn";
-interface Config {
-    project_id: string;
-    secret: string;
-    env: string;
-    timeout?: number;
-    agent?: http.Agent;
-}
-export declare class Client {
+import { BaseClient, ClientConfig } from "../shared/client";
+export declare class Client extends BaseClient {
     users: Users;
     magicLinks: MagicLinks;
     otps: OTPs;
@@ -26,7 +18,5 @@ export declare class Client {
     totps: TOTPs;
     webauthn: WebAuthn;
     cryptoWallets: CryptoWallets;
-    private fetchConfig;
-    constructor(config: Config);
+    constructor(config: ClientConfig);
 }
-export {};
